Finished Login System
User could now login to system

Upon the basic functionality of login there is also a validation if
user trying to access the login form after they already login. The
validation uses Jquery and check the user if they are already login
before loading the form. When they are already login they will be
redirected to the profile form

This also added to the registration and soon to all of form that don't
need to be access after logging in.